updating code with recommended changes
